+++ napari/layers/vectors/vectors_util.py
@@ -1,6 +1,6 @@
 import numpy as np
 from copy import copy
-from ...util import segment_normal
+from ..layer_utils import segment_normal
 
 
 def vectors_to_coordinates(vectors):
