+++ napari/util/tests/test_misc.py
@@ -12,13 +12,7 @@ from napari.util.misc import (
     get_pyramid_and_rgb,
     fast_pyramid,
     trim_pyramid,
-    calc_data_range,
     StringEnum,
-)
-
-
-data_dask = da.random.random(
-    size=(100_000, 1000, 1000), chunks=(1, 1000, 1000)
 )
 
 
@@ -34,63 +28,6 @@ def test_is_rgb():
 
     shape = (10, 15, 4)
     assert is_rgb(shape)
-
-
-def test_calc_data_range():
-    # all zeros should return [0, 1] by default
-    data = np.zeros((10, 10))
-    clim = calc_data_range(data)
-    assert np.all(clim == [0, 1])
-
-    # all ones should return [0, 1] by default
-    data = np.ones((10, 10))
-    clim = calc_data_range(data)
-    assert np.all(clim == [0, 1])
-
-    # return min and max
-    data = np.random.random((10, 15))
-    data[0, 0] = 0
-    data[0, 1] = 2
-    clim = calc_data_range(data)
-    assert np.all(clim == [0, 2])
-
-    # return min and max
-    data = np.random.random((6, 10, 15))
-    data[0, 0, 0] = 0
-    data[0, 0, 1] = 2
-    clim = calc_data_range(data)
-    assert np.all(clim == [0, 2])
-
-    # Try large data
-    data = np.zeros((1000, 2000))
-    data[0, 0] = 0
-    data[0, 1] = 2
-    clim = calc_data_range(data)
-    assert np.all(clim == [0, 2])
-
-    # Try large data mutlidimensional
-    data = np.zeros((3, 1000, 1000))
-    data[0, 0, 0] = 0
-    data[0, 0, 1] = 2
-    clim = calc_data_range(data)
-    assert np.all(clim == [0, 2])
-
-
-def test_calc_data_fast_uint8():
-    data = da.random.randint(
-        0,
-        100,
-        size=(100_000, 1000, 1000),
-        chunks=(1, 1000, 1000),
-        dtype=np.uint8,
-    )
-    assert calc_data_range(data) == [0, 255]
-
-
-@pytest.mark.timeout(2)
-def test_calc_data_range_fast_big():
-    val = calc_data_range(data_dask)
-    assert len(val) > 0
 
 
 def test_is_pyramid():
@@ -129,11 +66,6 @@ def test_is_pyramid():
     s = 8192
     data = [da.ones((s,) * 3), da.ones((s // 2,) * 3), da.ones((s // 4,) * 3)]
     assert is_pyramid(data)
-
-
-@pytest.mark.timeout(2)
-def test_timing_is_pyramid_big():
-    assert not is_pyramid(data_dask)
 
 
 def test_trim_pyramid():
